Share Mongo SRV reflection support
