+++ instrumentation/mongo/mongo-3.7/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/mongo/v3_7/MongoConfiguredTargetTest.java
@@ -7,19 +7,16 @@ package io.opentelemetry.javaagent.instrumentation.mongo.v3_7;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
-import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 import com.mongodb.MongoClientSettings;
 import com.mongodb.ServerAddress;
 import com.mongodb.UnixServerAddress;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
-import com.mongodb.connection.ClusterSettings;
 import io.opentelemetry.instrumentation.mongo.testing.AbstractMongoConfiguredTargetTest;
 import io.opentelemetry.instrumentation.mongo.testing.ClusterIdCapture;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
-import java.lang.reflect.Method;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -88,16 +85,12 @@ class MongoConfiguredTargetTest extends AbstractMongoConfiguredTargetTest {
 
   @Test
   void srvHostIsPreferredOverTheSeedsItStandsIn() {
-    // srvHost was added in 3.10
-    Method srvHost = srvHostSetter();
-    assumeTrue(srvHost != null);
-
     ClusterIdCapture clusterIdCapture = new ClusterIdCapture();
     MongoClientSettings settings =
         MongoClientSettings.builder()
             .applyToClusterSettings(
                 builder -> {
-                  invoke(srvHost, builder, "cluster0.example.invalid");
+                  applySrvHost(builder, "cluster0.example.invalid");
                   builder.addClusterListener(clusterIdCapture);
                 })
             .build();
@@ -129,21 +122,5 @@ class MongoConfiguredTargetTest extends AbstractMongoConfiguredTargetTest {
     }
 
     assertFindSpan(null, null);
-  }
-
-  private static Method srvHostSetter() {
-    try {
-      return ClusterSettings.Builder.class.getMethod("srvHost", String.class);
-    } catch (NoSuchMethodException ignored) {
-      return null;
-    }
-  }
-
-  private static void invoke(Method method, Object target, Object argument) {
-    try {
-      method.invoke(target, argument);
-    } catch (ReflectiveOperationException e) {
-      throw new IllegalStateException(e);
-    }
   }
 }
