+++ instrumentation/mongo/mongo-async-3.3/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/mongoasync/v3_3/MongoAsyncConfiguredTargetTest.java
@@ -5,8 +5,6 @@
 
 package io.opentelemetry.javaagent.instrumentation.mongoasync.v3_3;
 
-import static org.junit.jupiter.api.Assumptions.assumeTrue;
-
 import com.mongodb.ServerAddress;
 import com.mongodb.async.client.MongoClient;
 import com.mongodb.async.client.MongoClientSettings;
@@ -16,7 +14,6 @@ import io.opentelemetry.instrumentation.mongo.testing.AbstractMongoConfiguredTar
 import io.opentelemetry.instrumentation.mongo.testing.ClusterIdCapture;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
-import java.lang.reflect.Method;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -91,13 +88,9 @@ class MongoAsyncConfiguredTargetTest extends AbstractMongoConfiguredTargetTest {
 
   @Test
   void srvHostIsPreferredOverTheSeedsItStandsIn() {
-    // srvHost was added in 3.10
-    Method srvHost = srvHostSetter();
-    assumeTrue(srvHost != null);
-
     ClusterIdCapture clusterIdCapture = new ClusterIdCapture();
     ClusterSettings.Builder clusterSettings = ClusterSettings.builder();
-    invoke(srvHost, clusterSettings, "cluster0.example.invalid");
+    applySrvHost(clusterSettings, "cluster0.example.invalid");
     MongoClientSettings settings =
         MongoClientSettings.builder()
             .clusterSettings(clusterSettings.addClusterListener(clusterIdCapture).build())
@@ -107,21 +100,5 @@ class MongoAsyncConfiguredTargetTest extends AbstractMongoConfiguredTargetTest {
     runCommand(createClient(settings, clusterIdCapture));
 
     assertFindSpan("mongodb+srv://cluster0.example.invalid", null);
-  }
-
-  private static Method srvHostSetter() {
-    try {
-      return ClusterSettings.Builder.class.getMethod("srvHost", String.class);
-    } catch (NoSuchMethodException ignored) {
-      return null;
-    }
-  }
-
-  private static void invoke(Method method, Object target, Object argument) {
-    try {
-      method.invoke(target, argument);
-    } catch (ReflectiveOperationException e) {
-      throw new IllegalStateException(e);
-    }
   }
 }
