+++ instrumentation/mongo/mongo-common/testing/src/main/java/io/opentelemetry/instrumentation/mongo/testing/AbstractMongoConfiguredTargetTest.java
@@ -29,12 +29,14 @@ import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
 import com.mongodb.ConnectionString;
 import com.mongodb.ServerAddress;
 import com.mongodb.connection.ClusterId;
+import com.mongodb.connection.ClusterSettings;
 import com.mongodb.connection.ConnectionDescription;
 import com.mongodb.connection.ServerId;
 import com.mongodb.event.CommandListener;
 import com.mongodb.event.CommandStartedEvent;
 import com.mongodb.event.CommandSucceededEvent;
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
+import java.lang.reflect.Method;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.bson.BsonDocument;
@@ -80,6 +82,25 @@ public abstract class AbstractMongoConfiguredTargetTest {
         return connectionString;
       }
     };
+  }
+
+  protected static void applySrvHost(ClusterSettings.Builder builder, String host) {
+    Method srvHost = srvHostSetter();
+    assumeTrue(srvHost != null);
+    try {
+      srvHost.invoke(builder, host);
+    } catch (ReflectiveOperationException e) {
+      throw new IllegalStateException(e);
+    }
+  }
+
+  // srvHost was added in 3.10
+  private static Method srvHostSetter() {
+    try {
+      return ClusterSettings.Builder.class.getMethod("srvHost", String.class);
+    } catch (NoSuchMethodException ignored) {
+      return null;
+    }
   }
 
   @Test
